requirements.txt
streamlit
groq
coloredlogs
python-decouple